move UUID from stdlib/Random into Base
